Moved checksum calc and comp modules to '/etc/checksum'
Also renamed 'ChecksumUtils' to 'ChecksumCalculator', as currently its the only thing it does + It better separates concerns and turns it into a valid standalone module.
