+++ include/proto_nmea/etc/checksum/ChecksumCalculator.h
@@ -2,15 +2,15 @@
 // Copyright (c) 2019 Takpit. All rights reserved.
 //
 
-#ifndef PROTO_NMEA_CHECKSUMUTILS_H
-#define PROTO_NMEA_CHECKSUMUTILS_H
+#ifndef PROTO_NMEA_CHECKSUMCALCULATOR_H
+#define PROTO_NMEA_CHECKSUMCALCULATOR_H
 
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
-#include "HexUtils.h"
+#include "proto_nmea/utility/HexUtils.h"
 
 uint32_t calculateChecksum(const unsigned char *begin, const unsigned char *end);
 
@@ -18,4 +18,4 @@ uint32_t calculateChecksum(const unsigned char *begin, const unsigned char *end)
 }
 #endif
 
-#endif //PROTO_NMEA_CHECKSUMUTILS_H
+#endif //PROTO_NMEA_CHECKSUMCALCULATOR_H
